[robust_estimation] refactoring
splitting the function into another file
